+++ assets/webpack/src/js/chart.js
@@ -26,27 +26,47 @@ class Chart {
     this.listeners = {};
 
     this.visible = false;
-    this.graph = null;
     this.interval = null;
     this.error = null;
 
-    this.range = null;
-    this.zoomRange = null;
+    this.graph = {
+      // A lazily initialized reference to the Plotly graph element.
+      element: null,
+
+      // The data currently displayed in the graph, as returned by the storage,
+      // but slightly adjusted: hex bitmaps as integers, etc. Currently,
+      // keeping this data around is not strictly necessary, but it might be
+      // useful in the future to support incremental updates.
+      x: null,
+      y: null,
+
+      // The step of the data currently displayed in the graph, as returned by
+      // the storage. This is currently needed to estimate the optimal mode for
+      // the Plotly data depending on the available space (i.e., number of
+      // columns, zoom level, etc.).
+      step: null,
+
+      // The range in the X axis for the data currently contained in the graph,
+      // as returned by the storage, but slightly adjusted: right boundary
+      // adjusted one step back, etc.
+      range: null,
+
+      // The range for the X axis currently displayed in the graph, if zoomed.
+      zoomRange: null,
+    };
 
     observer.observe(this.container);
   }
 
   async init() {
     try {
-      // Fetch metric, adjusting the effective step if necessary.
-      const metric = await this.getMetric();
-
-      // Calculate & store the range for the X axis. This may change during zoom
-      // events, and we need to know the original range to reset it.
-      this.range = this.getPlotlyXAxisRange(metric);
+      // Fetch metric (adjusting the effective step if necessary, overriding the
+      // selected aggregator if the metric is a bitmap, etc.) and immediately
+      // process it to prepare the graph for the first render.
+      this.processMetric(await this.getMetric());
 
       // Render the graph for the first time.
-      this.graph = this.renderGraph(metric);
+      this.graph.element = this.renderGraph();
 
       // Since the chart is being initialized, it is visible, so we can start
       // the refresh loop to keep the chart up-to-date.
@@ -68,20 +88,18 @@ class Chart {
     // If the chart is not visible, there is no need to update it. Refresh loops
     // are canceled when the chart is not visible, but this is a safety check.
     // Also, manual refreshes would be pointless for invisible charts.
-    if (this.graph != null && this.visible) {
+    if (this.visible && this.graph.element != null) {
       try {
-        // Fetch metric, adjusting the effective step if necessary.
-        const metric = await this.getMetric();
-
-        // Calculate & store the range for the X axis. This may change during
-        // zoom events, and we need to know the original range to reset it.
-        this.range = this.getPlotlyXAxisRange(metric);
+        // Fetch metric (adjusting the effective step if necessary, overriding the
+        // selected aggregator if the metric is a bitmap, etc.) and immediately
+        // process it to prepare the graph for the update.
+        this.processMetric(await this.getMetric());
 
         // Update the graph with the new samples. For now we are just
         // re-rendering the whole graph, but in the future we should be able to
         // update it in a more efficient way. The server-side already supports
         // this kind of incremental updates.
-        this.updateGraph(metric);
+        this.updateGraph(false);
 
         // All good, clear any previous error.
         this.clearError();
@@ -95,23 +113,29 @@ class Chart {
   }
 
   handleGraphRelayout(event) {
+    // Update the zoom range of the graph.
     if (event['xaxis.range[0]'] && event['xaxis.range[1]']) {
-      this.zoomRange = [
+      this.graph.zoomRange = [
         new Date(event['xaxis.range[0]']),
         new Date(event['xaxis.range[1]']),
       ];
     } else if (event['xaxis.range'] && Array.isArray(event['xaxis.range']) && event['xaxis.range'].length === 2) {
-      this.zoomRange = [
+      this.graph.zoomRange = [
         new Date(event['xaxis.range'][0]),
         new Date(event['xaxis.range'][1]),
       ];
     } else {
-      this.zoomRange = null;
+      this.graph.zoomRange = null;
     }
 
+    // Force a re-render of the graph. The zoom range is already applied, but we
+    // might want to adjust other properties of the graph (e.g., the data mode).
+    this.updateGraph(true);
+
+    // Inform listeners about the zoom event.
     this.notifyEventListeners('zoom', {
       target: this,
-      range: this.zoomRange,
+      range: this.graph.zoomRange,
     });
   }
 
@@ -145,7 +169,7 @@ class Chart {
   setVisible(visible) {
     this.visible = visible;
     if (this.visible) {
-      if (this.graph == null) {
+      if (this.graph.element == null) {
         this.init();
       } else if (this.interval == null) {
         this.setupInterval();
@@ -156,6 +180,7 @@ class Chart {
   }
 
   redraw(filter, verbosity, columns) {
+    // Hide the chart if the metric name does not match the filter & verbosity.
     let hidden = verbosity === 'normal' && this.metric.debug;
     if (!hidden) {
       const terms = filter.split(/\s+/).filter(term => term.length > 0);
@@ -165,6 +190,7 @@ class Chart {
     }
     this.container.classList.toggle('d-none', hidden);
 
+    // Rearrange CSS classes to adjust to the number of columns.
     this.container.classList.forEach(className => {
       if (className.startsWith('col-')) {
         this.container.classList.remove(className);
@@ -172,8 +198,10 @@ class Chart {
     });
     this.container.classList.add(`col-${12 / columns}`);
 
-    if (this.graph != null) {
-      Plotly.Plots.resize(this.graph);
+    // If the chart has already been initialized (whether visible or not),
+    // update it to reflect the new size, the potential new data mode, etc.
+    if (this.graph.element != null) {
+      this.updateGraph(true);
     }
   }
 
@@ -200,20 +228,12 @@ class Chart {
   }
 
   setZoomRange(range) {
-    // Ignore if the range is the same as the current one.
-    if (this.zoomRange === range) {
-      return;
-    }
-    this.zoomRange = range;
+    this.graph.zoomRange = range;
 
-    // Adjust range in the X axis. It's important to apply the adjustment to
-    // every initialized chart, even if they are not visible.
-    if (this.graph != null) {
-      Plotly.update(this.graph, [], {
-        xaxis: {
-          range: this.zoomRange != null ? this.zoomRange : this.range,
-        },
-      });
+    // If the chart has already been initialized (whether visible or not),
+    // update it to reflect the new range, the potential new data mode, etc.
+    if (this.graph.element != null) {
+      this.updateGraph(true);
     }
   }
 
@@ -221,8 +241,8 @@ class Chart {
     observer.unobserve(this.container);
     this.stopInterval();
     this.clearError();
-    if (this.graph != null) {
-      Plotly.purge(this.graph);
+    if (this.graph.element != null) {
+      Plotly.purge(this.graph.element);
     }
   }
 
@@ -234,9 +254,9 @@ class Chart {
     this.stopInterval();
 
     if (this.visible) {
-      if (this.graph != null && this.refreshInterval > 0) {
+      if (this.graph.element != null && this.refreshInterval > 0) {
         this.interval = setInterval(this.handleRefresh.bind(this), this.refreshInterval * 1000);
-      } else if (this.graph == null) {
+      } else if (this.graph.element == null) {
         this.init();
       }
     }
@@ -331,15 +351,45 @@ class Chart {
     return Math.ceil(samples / maxSamples) * this.step;
   }
 
-  renderGraph(metric) {
+  processMetric(metric) {
+    // Prepare X & Y data for Plotly.
+    this.graph.x = [];
+    this.graph.y = [];
+    metric.samples.forEach(sample => {
+      this.graph.x.push(sample[0]);
+      // Bitmap metrics are returned as an hex string. For now, we represent the
+      // number of bits set to 1 in the bitmap as the Y value. This will be
+      // improved in the future using a different visualization for bitmap
+      // metrics.
+      this.graph.y.push(
+        this.metric.flag === 'b' ?
+          BigInt(`0x${sample[1]}`).toString(2).split('').filter(bit => bit === '1').length :
+          sample[1]);
+    });
+
+    // Store the step (already adjusted to be optimal for the space available)
+    // as returned by the storage.
+    this.graph.step = metric.step;
+
+    // Calculate & store the range for the X axis. This may change during
+    // zoom events, and we need to know the original range to reset it.
+    this.graph.range = [
+      metric.from,
+      new Date(metric.to.getTime() - metric.step * 1000),
+    ];
+  }
+
+  renderGraph() {
+    // Decide range to be used in the X axis.
+    const range = this.graph.zoomRange != null ? this.graph.zoomRange : this.graph.range;
+
     // Prepare data for Plotly.
-    const [x, y] = this.getPlotlyDataXY(metric);
     const data = [
       {
-        x: x,
-        y: y,
+        x: this.graph.x,
+        y: this.graph.y,
         type: 'scatter',
-        mode: this.estimatePlotlyDataMode(metric),
+        mode: this.estimatePlotlyDataMode(...range, this.graph.step),
         marker: { size: 4 },
         hovertemplate: '<b>X:</b> %{x|%Y-%m-%d %H:%M:%S}<br><b>Y:</b> %{y:,.1f}<extra></extra>',
         connectgaps: false,
@@ -359,7 +409,7 @@ class Chart {
       xaxis: {
         fixedrange: false,
         griddash: 'dash',
-        range: this.zoomRange != null ? this.zoomRange : this.range,
+        range: Array.from(range), // Beware the array needs to be cloned.
         autorange: false,
       },
       yaxis: {
@@ -401,55 +451,42 @@ class Chart {
     };
 
     // Render the graph.
-    const graph = this.container.querySelector('.graph');
-    Plotly.newPlot(graph, data, layout, config);
+    const graphElement = this.container.querySelector('.graph');
+    Plotly.newPlot(graphElement, data, layout, config);
 
     // Handle Plotly events.
-    graph.on('plotly_relayout', this.handleGraphRelayout.bind(this));
+    graphElement.on('plotly_relayout', this.handleGraphRelayout.bind(this));
 
     // Done!
-    return graph;
+    return graphElement;
   }
 
-  updateGraph(metric) {
+  updateGraph(sameData) {
+    // Decide range to be used in the X axis.
+    const range = this.graph.zoomRange != null ? this.graph.zoomRange : this.graph.range;
+
     // Prepare data for Plotly.
-    const [x, y] = this.getPlotlyDataXY(metric);
     const data = {
-      x: [x],
-      y: [y],
-      mode: this.estimatePlotlyDataMode(metric),
+      mode: this.estimatePlotlyDataMode(...range, this.graph.step),
     };
+    if (!sameData) {
+      data.x = [this.graph.x];
+      data.y = [this.graph.y];
+    }
 
     // Prepare layout for Plotly.
     const layout = {
       xaxis: {
-        range: this.zoomRange != null ? this.zoomRange : this.range,
+        range: Array.from(range), // Beware the array needs to be cloned.
       },
     };
 
     // Update the graph!
-    Plotly.update(this.graph, data, layout);
+    Plotly.update(this.graph.element, data, layout);
   }
 
-  getPlotlyDataXY(metric) {
-    const x = [];
-    const y = [];
-    metric.samples.forEach(sample => {
-      x.push(sample[0]);
-      // Bitmap metrics are returned as an hex string. For now, we represent the
-      // number of bits set to 1 in the bitmap as the Y value. This will be
-      // improved in the future using a different visualization for bitmap
-      // metrics.
-      y.push(
-        this.metric.flag === 'b' ?
-          BigInt(`0x${sample[1]}`).toString(2).split('').filter(bit => bit === '1').length :
-          sample[1]);
-    });
-    return [x, y];
-  }
-
-  estimatePlotlyDataMode(metric) {
-    const samples = (helpers.dateToUnix(metric.to) - helpers.dateToUnix(metric.from)) / metric.step;
+  estimatePlotlyDataMode(from, to, step) {
+    const samples = (helpers.dateToUnix(to) - helpers.dateToUnix(from)) / step;
     const containerWidth = this.container.clientWidth;
     const minSpacing = 6;
     const maxSamples = Math.floor(0.9 * containerWidth / minSpacing);
@@ -457,10 +494,6 @@ class Chart {
       return 'lines';
     }
     return 'lines+markers';
-  }
-
-  getPlotlyXAxisRange(metric) {
-    return [metric.from, new Date(metric.to.getTime() - metric.step * 1000)];
   }
 }