+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 - 0.3.0-1 (2025-01-26):
     + Added extra logging during bootstrap / rotation of the storage.
-    + Adjusted shape and with of lines in charts.
+    + Adjusted shape and width of lines in charts.
     + Added support to zoom in and out in the charts.
     + Increased default DuckDB memory limit to 512 MiB.
     + Added support to filter metrics by multiple search terms.