+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 - ?
     + Reworked the internals of the chart widget.
     + Modified the behavior to update (but not apply) the time range picker during zoom in/out events.
+    + Limited the zoom of charts to a minimum and maximum range.
+    + Enabled zooming in and out of charts using the mouse wheel.
+    + Adjusted the width of lines in charts.
 
 - 0.3.1-1 (2025-01-26):
     + Fixed filtering of metrics when no search terms are provided.