+++ CHANGELOG.md
@@ -2,10 +2,10 @@
     + Reworked the internals of the chart widget.
     + Modified the behavior to update (but not apply) the time range picker during zoom in/out events.
     + Limited the zoom of charts to a minimum and maximum range.
-    + Enabled zooming in and out of charts using the mouse wheel.
     + Adjusted the width of lines in charts.
     + Added a filter history to easily reuse previous filter strings.
     + Added visual feedback to charts when the effective step is different from the selected step.
+    + Improved refreshing of charts.
 
 - 0.3.1-1 (2025-01-26):
     + Fixed filtering of metrics when no search terms are provided.