+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-- ?
+- 0.4.0-1 (2025-01-28):
     + Reworked the internals of the chart widget.
     + Modified the behavior to update (but not apply) the time range picker during zoom in/out events.
     + Limited the zoom of charts to a minimum and maximum range.